+++ Assets/Core/Runtime/DataBind/GameObjectArrayBind.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+namespace KKSFramework.DataBind
+{
+    public class GameObjectArrayBind : BindableComponents<GameObject>
+    {
+    }
+}