+++ Assets/Core/Runtime/DataBind/TextArrayBind.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine.UI;
-
-namespace KKSFramework.DataBind
-{
-    public class TextArrayBind : BindableComponents<Text>
-    {
-    }
-}