+++ Assets/Test/SomePage.cs
@@ -1,54 +1,57 @@
+using System.Linq;
+using KKSFramework;
+using KKSFramework.DataBind;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace KKSFramework.DataBind
+public class SomePage : MonoBehaviour, IResolveTarget
 {
-    public class SomePage : MonoBehaviour, IBinder
-    {
-        #region Fields & Property
+    #region Fields & Property
+
+    public Context context;
 
 #pragma warning disable CS0649
 
-        [ResolveUI("ScoreText")]
-        private Text _scoreText;
+    [Resolver]
+    private Text _scoreText;
 
-        [ResolveUI("ScoreTexts")]
-        private Text[] _texts;
+    [Resolver ("ScoreTexts")]
+    private Text[] _texts;
 
-        [ResolveUI("Images")]
-        private Image[] _images;
+    private Button[] _textButtons;
 
-        [ResolveUI("Button")]
-        private Button _button;
-        
+    [Resolver ("Elements")]
+    private SomeElement[] _someElement;
+
 #pragma warning restore CS0649
 
-        #endregion
+    #endregion
 
+
+    #region UnityMethods
+
+    private void Start ()
+    {
+        _textButtons = context.Resolve<GameObject[]> ("ScoreTexts").Select (x => x.GetComponent<Button> ()).ToArray ();
         
-        #region UnityMethods
-
-        private void Start ()
+        _scoreText.text = "Good";
+        _texts[0].text = "haha";
+        _textButtons[0].onClick.AddListener (() =>
         {
-            _scoreText.text = "Good";
-            _texts[0].text = "haha";
-            _button.onClick.AddListener (() =>
-            {
-                _images[0].color = Color.red;
-            });
-            Debug.Log (_texts.Length);
-        }
-
-        #endregion
-
-
-        #region Methods
-
-        #endregion
-
-
-        #region EventMethods
-
-        #endregion
+            _textButtons[0].gameObject.SetActive (false);
+        });
+        _someElement.Foreach (x => x.Debug ());
     }
+
+    #endregion
+
+
+    #region Methods
+
+    #endregion
+
+
+    #region EventMethods
+
+    #endregion
 }